refactor(report): 금액·카운트 응답 필드 Long을 Integer로 정리
프론트 JSON 정밀도 우려(JS 안전 정수 2^53-1 초과 시 손실)에서 시작해, report·chat·user·
adjuster 도메인의 금액·카운트 필드를 Integer로 통일했다. QueryDSL count()·Spring Data
Page·Redis Lua 스크립트처럼 프레임워크가 Long을 강제하는 내부 계층(Row·RepositoryImpl)은
그대로 두고, 컨트롤러가 실제로 내려주는 응답 DTO 경계에서만 좁혔다. 금액 컬럼은 V32
마이그레이션으로 bigint에서 integer로 맞췄다.

Closes #195

+++ src/main/java/com/soma/backend/domain/chat/dto/ChatMessageResponse.java
@@ -23,6 +23,6 @@ public record ChatMessageResponse(
     LocalDateTime createdAt) {
 
   /** 첨부 표시 정보 + 단기 presigned GET URL. */
-  public record Attachment(String url, String name, String contentType, Long size) {
+  public record Attachment(String url, String name, String contentType, Integer size) {
   }
 }
